+++ hypertunity/__init__.py
@@ -2,3 +2,4 @@ from .optimisation import *
 from .reports import *
 from .scheduling import *
 from .trial import *
+from .domain import *