+++ hypertunity/optimisation/__init__.py
@@ -1,5 +1,4 @@
 from .base import *
 from .bo import *
-from .domain import *
 from .exhaustive import *
 from .random import *