+++ hypertunity/optimisation/exhaustive.py
@@ -2,8 +2,8 @@
 
 from typing import List
 
+from hypertunity.domain import Domain, Sample, DomainNotIterableError
 from hypertunity.optimisation.base import Optimiser
-from hypertunity.optimisation.domain import Domain, Sample, DomainNotIterableError
 
 __all__ = [
     "GridSearch",