+++ hypertunity/scheduling/tests/test_scheduler.py
@@ -3,14 +3,16 @@ import tempfile
 
 import pytest
 
-import hypertunity as ht
+from hypertunity.domain import Domain, Sample
+from hypertunity.optimisation import base
+
 from . import script
 from ..jobs import Job, SlurmJob
 from ..scheduler import Scheduler
 
 
-def square(sample: ht.Sample) -> ht.EvaluationScore:
-    return ht.EvaluationScore(sample["x"] ** 2)
+def square(sample: Sample) -> base.EvaluationScore:
+    return base.EvaluationScore(sample["x"] ** 2)
 
 
 def run_jobs(jobs):
@@ -24,7 +26,7 @@ def run_jobs(jobs):
 
 @pytest.mark.timeout(10.0)
 def test_local_from_script_and_function():
-    domain = ht.Domain({"x": {0, 1, 2, 3}, "y": [-1., 1.], "z": {"123", "abc"}}, seed=7)
+    domain = Domain({"x": {0, 1, 2, 3}, "y": [-1., 1.], "z": {"123", "abc"}}, seed=7)
     jobs = [Job(task="hypertunity/scheduling/tests/script.py::main",
                 args=(*domain.sample().as_namedtuple(),)) for _ in range(10)]
     results = run_jobs(jobs)
@@ -33,7 +35,7 @@ def test_local_from_script_and_function():
 
 @pytest.mark.timeout(10.0)
 def test_local_from_script_and_cmdline_args():
-    domain = ht.Domain({"x": {0, 1, 2, 3}, "y": [-1., 1.], "z": {"123", "abc"}}, seed=7)
+    domain = Domain({"x": {0, 1, 2, 3}, "y": [-1., 1.], "z": {"123", "abc"}}, seed=7)
     jobs = [Job(task="hypertunity/scheduling/tests/script.py",
                 args=(*domain.sample().as_namedtuple(),),
                 meta={"binary": "python"}) for _ in range(10)]
@@ -43,7 +45,7 @@ def test_local_from_script_and_cmdline_args():
 
 @pytest.mark.timeout(10.0)
 def test_local_from_script_and_cmdline_named_args():
-    domain = ht.Domain({"--x": {0, 1, 2, 3}, "--y": [-1., 1.], "--z": {"acb123", "abc"}}, seed=7)
+    domain = Domain({"--x": {0, 1, 2, 3}, "--y": [-1., 1.], "--z": {"acb123", "abc"}}, seed=7)
     jobs = [Job(task="hypertunity/scheduling/tests/script.py",
                 args=domain.sample().as_dict(),
                 meta={"binary": "python"}) for _ in range(10)]
@@ -53,7 +55,7 @@ def test_local_from_script_and_cmdline_named_args():
 
 @pytest.mark.timeout(10.0)
 def test_local_from_fn():
-    domain = ht.Domain({"x": [0., 1.]}, seed=7)
+    domain = Domain({"x": [0., 1.]}, seed=7)
     jobs = [Job(task=square, args=(domain.sample(),)) for _ in range(10)]
     results = run_jobs(jobs)
     assert all([r.data.value == square(*j.args).value for r, j in zip(results, jobs)])
@@ -62,7 +64,7 @@ def test_local_from_fn():
 @pytest.mark.slurm
 @pytest.mark.timeout(60.0)
 def test_slurm_from_script():
-    domain = ht.Domain({"x": {0, 1, 2, 3}, "y": [-1., 1.], "z": {"123", "abc"}}, seed=7)
+    domain = Domain({"x": {0, 1, 2, 3}, "y": [-1., 1.], "z": {"123", "abc"}}, seed=7)
     jobs, dirs = [], []
     n_jobs = 4
     for i in range(n_jobs):